requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib[bcrypt]
pydantic[email]
python-jose
python-dotenv